Use cligj for plugin registration for #228 and #224.
