+++ tests/test_fio_load.py
@@ -1,21 +1,19 @@
-import json
 import os
 import tempfile
 
-import click
 from click.testing import CliRunner
 
 import fiona
-from fiona.fio import fio
+from fiona.fio.main import main_group
 
 from .fixtures import (
-    feature_collection, feature_collection_pp, feature_seq, feature_seq_pp_rs)
+    feature_collection, feature_seq, feature_seq_pp_rs)
 
 
 def test_err():
     runner = CliRunner()
     result = runner.invoke(
-        fio.load, [], '', catch_exceptions=False)
+        main_group, ['load'], '', catch_exceptions=False)
     assert result.exit_code == 2
 
 
@@ -27,7 +25,7 @@ def test_exception(tmpdir=None):
         tmpfile = str(tmpdir.join('test.shp'))
     runner = CliRunner()
     result = runner.invoke(
-        fio.load, ['-f', 'Shapefile', tmpfile], '42', catch_exceptions=False)
+        main_group, ['load', '-f', 'Shapefile', tmpfile], '42', catch_exceptions=False)
     assert result.exit_code == 1
 
 
@@ -39,7 +37,7 @@ def test_collection(tmpdir=None):
         tmpfile = str(tmpdir.join('test.shp'))
     runner = CliRunner()
     result = runner.invoke(
-        fio.load, ['-f', 'Shapefile', tmpfile], feature_collection)
+        main_group, ['load', '-f', 'Shapefile', tmpfile], feature_collection)
     assert result.exit_code == 0
     assert len(fiona.open(tmpfile)) == 2
 
@@ -52,7 +50,7 @@ def test_seq_rs(tmpdir=None):
         tmpfile = str(tmpdir.join('test.shp'))
     runner = CliRunner()
     result = runner.invoke(
-        fio.load, ['-f', 'Shapefile', tmpfile], feature_seq_pp_rs)
+        main_group, ['load', '-f', 'Shapefile', tmpfile], feature_seq_pp_rs)
     assert result.exit_code == 0
     assert len(fiona.open(tmpfile)) == 2
 
@@ -65,6 +63,6 @@ def test_seq_no_rs(tmpdir=None):
         tmpfile = str(tmpdir.join('test.shp'))
     runner = CliRunner()
     result = runner.invoke(
-        fio.load, ['-f', 'Shapefile', '--sequence', tmpfile], feature_seq)
+        main_group, ['load', '-f', 'Shapefile', '--sequence', tmpfile], feature_seq)
     assert result.exit_code == 0
     assert len(fiona.open(tmpfile)) == 2
